Flask-API/requirements.txt
Flask
Gunicorn
google-generativeai
python-dotenv
youtube-transcript-api
Flask-Cors
